api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
tensorflow==2.20.0
pillow==10.1.0
numpy==1.26.2
requests==2.31.0
python-multipart==0.0.9
